+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 * Initial release - Supports Radar Chart with lots of customizations.
 
-##[0.0.2] - 01/09/2019
+## [0.0.2] - 01/09/2019
 
 * Fixed screenshots
+
+## [0.1.0] - 21/09/2019
+
+* Added **labelColor** parameter to give the desired color to labels, default value is **Colors.white**
+* Added a check for **values** (Minimum 3 values are required for plotting Radar Chart), else it will throw **ArgumentError** 
+